DRCInfo->Board Version to one line
After the merge, the additional information in DRH INFO causes the text to no longer fit on the screen. To address this, "Board Version" has been modified to display the information in a single line.

Ideally, everything should fit on the screen, but a scrolling system will need to be added when the content exceeds the available space.